refactor current assets investments
