+++ src/main/java/uk/gov/companieshouse/api/accounts/model/rest/smallfull/notes/CurrentAssetsInvestments.java
@@ -1,9 +1,9 @@
-package uk.gov.companieshouse.api.accounts.model.rest.notes;
+package uk.gov.companieshouse.api.accounts.model.rest.smallfull.notes;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.google.gson.Gson;
-import uk.gov.companieshouse.api.accounts.model.rest.RestObject;
+import uk.gov.companieshouse.api.accounts.model.rest.Note;
 import uk.gov.companieshouse.api.accounts.validation.CharSetValid;
 import uk.gov.companieshouse.charset.CharSet;
 
@@ -12,7 +12,7 @@ import javax.validation.constraints.Size;
 import java.util.Objects;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
-public class CurrentAssetsInvestments extends RestObject {
+public class CurrentAssetsInvestments extends Note {
 
     private static final int MAX_FIELD_LENGTH = 20000;
     private static final int MIN_FIELD_LENGTH = 1;
