+++ src/main/java/uk/gov/companieshouse/api/accounts/validation/CurrentAssetsInvestmentsValidator.java
@@ -2,20 +2,21 @@ package uk.gov.companieshouse.api.accounts.validation;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+import uk.gov.companieshouse.api.accounts.enumeration.AccountingNoteType;
 import uk.gov.companieshouse.api.accounts.exception.DataException;
 import uk.gov.companieshouse.api.accounts.model.rest.BalanceSheet;
 import uk.gov.companieshouse.api.accounts.model.rest.CurrentPeriod;
-import uk.gov.companieshouse.api.accounts.model.rest.notes.CurrentAssetsInvestments;
+import uk.gov.companieshouse.api.accounts.model.rest.smallfull.notes.CurrentAssetsInvestments;
 import uk.gov.companieshouse.api.accounts.model.validation.Errors;
 import uk.gov.companieshouse.api.accounts.service.impl.CurrentPeriodService;
 import uk.gov.companieshouse.api.accounts.service.impl.PreviousPeriodService;
 import uk.gov.companieshouse.api.accounts.service.response.ResponseObject;
+import uk.gov.companieshouse.api.model.transaction.Transaction;
 
 import javax.servlet.http.HttpServletRequest;
-import javax.validation.Valid;
 
 @Component
-public class CurrentAssetsInvestmentsValidator extends BaseValidator {
+public class CurrentAssetsInvestmentsValidator extends BaseValidator implements NoteValidator<CurrentAssetsInvestments> {
 
     private static final String CURRENT_ASSETS_DETAILS_PATH = "$.current_assets_investments.details";
 
@@ -29,9 +30,10 @@ public class CurrentAssetsInvestmentsValidator extends BaseValidator {
         this.previousPeriodService = previousPeriodService;
     }
 
-    public Errors validateCurrentAssetsInvestments(@Valid HttpServletRequest request,
-        CurrentAssetsInvestments currentAssetsInvestments,
-        String companyAccountsId) throws DataException {
+
+    @Override
+    public Errors validateSubmission(CurrentAssetsInvestments currentAssetsInvestments, Transaction transaction,
+                                     String companyAccountsId, HttpServletRequest request) throws DataException {
 
         Errors errors = new Errors();
 
@@ -98,4 +100,10 @@ public class CurrentAssetsInvestmentsValidator extends BaseValidator {
             return null;
         }
     }
+
+    @Override
+    public AccountingNoteType getAccountingNoteType() {
+        return AccountingNoteType.SMALL_FULL_CURRENT_ASSETS_INVESTMENTS;
+    }
+
 }
