+++ src/main/resources/notes.yml
@@ -5,3 +5,4 @@ notes:
     - stocks
     - employees
     - debtors
+    - current-assets-investments
